Remove auto-upgrade logic
This functionality breaks existing scripts.

I use vdesktop as part of an automated taskflow. When vdesktop prompts for an upgrade answer, it breaks the workflow. This is unexpected, as a script that worked on Monday shouldn't suddenly break on Tuesday because of a change on a remote Github server.

Additionally, the previous implementation wasn't very safe. If the user had locally modified files in their repository, vdesktop would blindly erase them without prompting for confirmation.

Instead of attempting an auto-upgrade, this makes vdesktop simply print upgrade instructions.